In some situations new date handling could cause a page to not display or to show strange chars together with the date. This behaviour has not been explained so far as all function callings seem to be fine. The problem has been narrowed to a date( $format, $stamp ); calling. $format and $stamp have been checked and are both OK. It looks like an internal PHP problem, but I can't be sure.
I've observed this problem in PHP 4.0.4. PHP 4.0.6 is working just fine and
don't show that strange behaviour.

In order to avoid that problem, I've observed that placing some chars
leading and trailing the date() format string avoids the problem (so maybe
an internal pointer to the string problem ?). So do I and I remove then once
the conversion is finished.

If someone can reproduce the bug with prior version and find a solution
please tell me.